backend/requirements.txt
black==19.10b0
fastapi==0.88.0
fastcache==1.1.0
mypy-extensions==0.4.3
param==1.12.0
pathspec==0.7.0
pyairtable==1.3.0
pydantic==1.10.2
pytest==6.2.4
python-dotenv==0.21.0
typing_extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.14.1


